Movingconversation picker into own service
